+++ lib/src/dialogs/actions.dart
@@ -1,45 +1,150 @@
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 
-import 'base_dialog.dart';
+import 'platform_widget.dart';
 
-class BasicDialogAction extends BaseDialog<TextButton, CupertinoDialogAction> {
-  const BasicDialogAction({
+class MaterialActionData {
+  MaterialActionData({
+    this.widgetKey,
+    this.onLongPress,
+    this.onHover,
+    this.onFocusChange,
+    this.style,
+    this.focusNode,
+    this.autofocus = false,
+    this.statesController,
+    this.icon,
+    this.useTextButton = true,
+  });
+  final Key? widgetKey;
+  final VoidCallback? onLongPress;
+  final void Function(bool)? onHover;
+  final void Function(bool)? onFocusChange;
+  final ButtonStyle? style;
+  final FocusNode? focusNode;
+  final bool autofocus;
+  final Clip clipBehavior = Clip.none;
+  final MaterialStatesController? statesController;
+  final Widget? icon;
+  final bool useTextButton;
+}
+
+class CupertinoActionData {
+  final Key? widgetKey;
+  final bool isDefaultAction;
+  final bool isDestructiveAction;
+  final TextStyle? textStyle;
+
+  CupertinoActionData({
+    this.widgetKey,
+    this.isDefaultAction = false,
+    this.isDestructiveAction = false,
+    this.textStyle,
+  });
+}
+
+class PlatformDialogAction
+    extends PlatformWidget<Widget, CupertinoDialogAction> {
+  const PlatformDialogAction({
     super.key,
-    super.useMaterial,
-    required this.title,
-    required this.onPressed,
-    this.isIosDestructiveAction = false,
-    this.isIosDefaultAction = false,
-    this.buttonStyle,
+    super.useMaterial = false,
+    required this.child,
+    this.onPressed,
+    this.material,
+    this.cupertino,
   });
 
-  final Widget title;
-  final VoidCallback onPressed;
+  /// child widget
+  final Widget child;
 
-  /// only for ios
-  final bool isIosDestructiveAction;
-  final bool isIosDefaultAction;
+  /// action onPressed callback
+  final VoidCallback? onPressed;
 
-  /// only for android
-  final ButtonStyle? buttonStyle;
+  /// action properties for material
+  final MaterialActionData? material;
+
+  /// action properties for cupertino
+  final CupertinoActionData? cupertino;
 
   @override
   CupertinoDialogAction buildCupertinoWidget(BuildContext context) {
     return CupertinoDialogAction(
+      key: cupertino?.widgetKey,
       onPressed: onPressed,
-      isDefaultAction: isIosDefaultAction,
-      isDestructiveAction: isIosDestructiveAction,
-      child: title,
+      isDefaultAction: cupertino?.isDefaultAction ?? false,
+      isDestructiveAction: cupertino?.isDestructiveAction ?? false,
+      textStyle: cupertino?.textStyle,
+      child: child,
     );
   }
 
   @override
-  TextButton buildMaterialWidget(BuildContext context) {
-    return TextButton(
-      style: buttonStyle,
+  Widget buildMaterialWidget(BuildContext context) {
+    final icon = material?.icon;
+    final useTextButton = material?.useTextButton ?? true;
+
+    if (icon != null && useTextButton) {
+      return TextButton.icon(
+        key: material?.widgetKey,
+        onPressed: onPressed,
+        onLongPress: material?.onLongPress,
+        onHover: material?.onHover,
+        onFocusChange: material?.onFocusChange,
+        style: material?.style,
+        focusNode: material?.focusNode,
+        autofocus: material?.autofocus ?? false,
+        clipBehavior: material?.clipBehavior ?? Clip.none,
+        statesController: material?.statesController,
+        icon: icon,
+        label: child,
+      );
+    }
+
+    if (icon != null && !useTextButton) {
+      return FilledButton.icon(
+        key: material?.widgetKey,
+        onPressed: onPressed,
+        onLongPress: material?.onLongPress,
+        onHover: material?.onHover,
+        onFocusChange: material?.onFocusChange,
+        style: material?.style,
+        focusNode: material?.focusNode,
+        autofocus: material?.autofocus ?? false,
+        clipBehavior: material?.clipBehavior ?? Clip.none,
+        statesController: material?.statesController,
+        icon: icon,
+        label: child,
+      );
+    }
+
+    if (useTextButton) {
+      return TextButton(
+        key: material?.widgetKey,
+        onPressed: onPressed,
+        onLongPress: material?.onLongPress,
+        onHover: material?.onHover,
+        onFocusChange: material?.onFocusChange,
+        style: material?.style,
+        focusNode: material?.focusNode,
+        autofocus: material?.autofocus ?? false,
+        clipBehavior: material?.clipBehavior ?? Clip.none,
+        statesController: material?.statesController,
+        child: child,
+      );
+    }
+
+    return FilledButton(
+      key: material?.widgetKey,
       onPressed: onPressed,
-      child: title,
+      onLongPress: material?.onLongPress,
+      onHover: material?.onHover,
+      onFocusChange: material?.onFocusChange,
+      style: material?.style,
+      focusNode: material?.focusNode,
+      autofocus: material?.autofocus ?? false,
+      clipBehavior: material?.clipBehavior ?? Clip.none,
+      statesController: material?.statesController,
+      child: child,
     );
   }
 }