+++ lib/src/dialogs/dialogs.dart
@@ -1,4 +1,6 @@
 export 'actions.dart';
-export 'base_dialog.dart';
-export 'data.dart';
+export 'basic_alert_dialog.dart';
+export 'basic_bottom_sheet.dart';
+export 'platform_bottom_sheet.dart';
 export 'platform_dialog.dart';
+export 'platform_widget.dart';